moved all split structs to split module
